+++ src/Presentation/SmartStore.Web/Administration/Controllers/ReturnRequestController.cs
@@ -156,17 +156,20 @@ namespace SmartStore.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageReturnRequests))
                 return AccessDeniedView();
 
-            var returnRequests = new List<ReturnRequestModel>();
-            foreach (var rr in _orderService.SearchReturnRequests(0, 0, 0, null).PagedForCommand(command))
+			var data = new List<ReturnRequestModel>();
+			var returnRequests = _orderService.SearchReturnRequests(0, 0, 0, null, command.Page - 1, command.PageSize);
+
+            foreach (var rr in returnRequests)
             {
                 var m = new ReturnRequestModel();
                 if (PrepareReturnRequestModel(m, rr, false))
-                    returnRequests.Add(m);
+                    data.Add(m);
             }
+
             var gridModel = new GridModel<ReturnRequestModel>
             {
-                Data = returnRequests,
-                Total = returnRequests.Count,
+                Data = data,
+                Total = returnRequests.TotalCount,
             };
             return new JsonResult
             {