+++ changelog.md
@@ -64,6 +64,7 @@
 * #405 Billiger feed: Wrong base price exported
 * #437 Mobile devices: Cannot add a product to the cart when it is grouped
 * PayPal Standard: Costs for checkout attributes were double charged
+* Paging of return request grid did not work
 
 
 ##SmartStore.NET 2.0.2#